Crate strife

A lightweight library for the Discord API.

Modules

client
The Client contains information about a single bots token and event handlers. Dispatching of events to handlers and starting shard connections are handled by the client.

constants
A collection of constants used by the library.

http
The HTTP module provides low-level functions for interacting with the Discord API.

model
Models of objects received from the Discord API.

Structs

Client
A client that connects to Discord via the WebSocket and API.

Http
An HTTP client for performing requests to the REST API.

Enums

Error
The common error enum returned by library functions within a Result.

Type Definitions

Result
The common result type returned by library functions.