Module strife::constants

A collection of constants used by the library.

Modules

close_codes
Gateway close event codes.

Enums

OpCode
Gateway opcodes.

Constants

EMBED_MAX_LENGTH
The maximum length of textual size of an embed message.

GATEWAY_VERSION
The gateway version used by the library, URI is retrieved via the REST API.

MESSAGE_MAX_LENGTH
The maximum length of a message in Unicode code points allowed by Discord.

USER_AGENT
The UserAgent header sent with every request.