Module strife::constants::close_codes

Gateway close event codes.

Constants

ALREADY_AUTHENTICATED
More than one identify payload was sent.

AUTHENTICATION_FAILED
The account token sent with your identify payload is incorrect.

DECODE_ERROR
An invalid payload was sent.

INVALID_SEQUENCE
The sequence sent when resuming the session was invalid.

INVALID_SHARD
An invalid shard was sent when identifying.

NOT_AUTHENTICATED
A payload was sent prior to identifying.

RATE_LIMITED
Payloads were sent too quickly.

SESSION_TIMEOUT
Session timed out.

SHARDING_REQUIRED
The session would have handled too many guilds.

UNKNOWN_ERROR
Unknown error. Try reconnecting?

UNKNOWN_OPCODE
An invalid gateway opcode or an invalid payload for an opcode was sent.