Struct strife::model::channel::message::Message

#[non_exhaustive]
pub struct Message {
    pub id: MessageId,
    pub channel_id: ChannelId,
    pub guild_id: Option<GuildId>,
    pub author: User,
    pub member: Option<PartialMember>,
    pub content: String,
    pub timestamp: DateTime<FixedOffset>,
    pub edited_timestamp: Option<DateTime<FixedOffset>>,
    pub tts: bool,
    pub mention_everyone: bool,
    pub mentions: Vec<MentionedUser>,
    pub mention_roles: Vec<RoleId>,
    pub mention_channels: Vec<MentionedChannel>,
    pub attachments: Vec<Attachment>,
    pub embeds: Vec<Embed>,
    pub reactions: Vec<Reaction>,
    pub pinned: bool,
    pub webhook_id: Option<WebhookId>,
    pub kind: MessageType,
    pub activity: Option<MessageActivity>,
    pub application: Option<MessageApplication>,
    pub message_reference: Option<MessageReference>,
    pub flags: MessageFlags,
}

A message sent in a text channel.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
id: MessageId

The ID of the message.

channel_id: ChannelId

The ID of the Channel the message was sent in.

guild_id: Option<GuildId>

The ID of the Guild the message was sent in.

author: User

The author of the message.

May not be a valid user, see notes for more information.

Notes

The author object follows the structure of the user object, but is only a valid user in the case where the message is generated by a user or bot user. If the message is generated by a webhook, the author object corresponds to the webhook's id, username, and avatar. You can tell if a message is generated by a webhook by checking for the webhook_id on the message object.

member: Option<PartialMember>

Partial guild member properties for the author, if the message was sent in a guild.

content: String

The content of the message.

timestamp: DateTime<FixedOffset>

When the message was sent.

edited_timestamp: Option<DateTime<FixedOffset>>

When the message was edited, if the message has been edited.

tts: bool

Whether the message was sent as a TTS (Text-To-Speech) message.

mention_everyone: bool

Whether the message mentions everyone.

mentions: Vec<MentionedUser>

The users specifically mentioned in the message.

mention_roles: Vec<RoleId>

The roles specifically mentioned in the message.

mention_channels: Vec<MentionedChannel>

The channels specifically mentioned in the message.

Only textual channels that are visible to everyone in a lurkable guild will ever be included.

attachments: Vec<Attachment>

The files attached to the message.

embeds: Vec<Embed>

The embedded content of the message.

reactions: Vec<Reaction>

The reactions to the message.

pinned: bool

Whether the message is pinned.

webhook_id: Option<WebhookId>

The webhook ID that generated the message, if the message was generated by a webhook.

kind: MessageType

The type of message.

activity: Option<MessageActivity>

The Rich Presence activity information sent with related embeds.

application: Option<MessageApplication>

The Rich Presence application information sent with related embeds.

message_reference: Option<MessageReference>

The reference data sent with a crossposted message.

flags: MessageFlags

Flags describing extra features of the message.