Module strife::model::guild

Models related to guilds.

Structs

CustomEmoji
A custom guild emoji with partial information.

Emoji
A guild emoji.

Guild
A guild in Discord represents an isolated collection of users and channels, and are often referred to as "servers" in the UI.

Member
A member of a guild.

PartialGuild
A guild with partial information.

PartialMember
A member of a guild, with partial information.

Role
Represents a set of permissions attached to a group of users.

Enums

AuditLogEvent
The type of action that occurred in an [AuditLogEntry].

ExplicitContentFilterLevel
The level of filter to apply to users that send messages containing explicit content.

GuildFeature
A feature enabled for a guild.

MessageNotificationLevel
The default level of message notifications in a guild.

MfaLevel
The required level of multi-factor authentication required for a guild.

PartialEmoji
An emoji, with partial information.

PremiumTier
The tier of premium for a guild, provided by Nitro boosts.

VerificationLevel
The required level of criteria a user must meet, prior to being able to send messages in a Guild.