Struct strife::model::guild::Guild

#[non_exhaustive]
pub struct Guild {
    pub splash: Option<String>,
    pub owner_id: UserId,
    pub region: VoiceRegionId,
    pub afk_channel_id: Option<ChannelId>,
    pub afk_timeout: u64,
    pub embed_enabled: bool,
    pub embed_channel_id: Option<ChannelId>,
    pub verification_level: VerificationLevel,
    pub message_notifications: MessageNotificationLevel,
    pub explicit_content_filter: ExplicitContentFilterLevel,
    pub roles: HashMap<RoleId, Role>,
    pub emojis: HashMap<EmojiId, Emoji>,
    pub features: HashSet<GuildFeature>,
    pub mfa_level: MfaLevel,
    pub application_id: Option<ApplicationId>,
    pub widget_enabled: bool,
    pub widget_channel_id: Option<ChannelId>,
    pub system_channel_id: Option<ChannelId>,
    pub joined_at: Option<DateTime<FixedOffset>>,
    pub large: bool,
    pub unavailable: bool,
    pub member_count: Option<u64>,
    pub voice_states: Vec<VoiceState>,
    pub members: HashMap<UserId, Member>,
    pub channels: HashMap<ChannelId, GuildChannel>,
    pub presences: HashMap<UserId, Presence>,
    pub max_members: Option<u64>,
    pub vanity_url_code: Option<String>,
    pub description: Option<String>,
    pub banner: Option<String>,
    pub premium_tier: PremiumTier,
    pub boost_count: u64,
    pub preferred_locale: Locale,
    // some fields omitted
}

A guild in Discord represents an isolated collection of users and channels, and are often referred to as "servers" in the UI.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
splash: Option<String>

The hash of the guild splash.

owner_id: UserId

The ID of the owner of the guild.

region: VoiceRegionId

The ID of the guild voice region.

afk_channel_id: Option<ChannelId>

The ID of the AFK channel.

afk_timeout: u64

The AFK timeout in seconds.

embed_enabled: bool

Whether the guild is embeddable (eg. widget).

embed_channel_id: Option<ChannelId>

The ID of the channel that the embed widget will generate in invite to.

verification_level: VerificationLevel

The verification level required for the guild.

message_notifications: MessageNotificationLevel

The default message notification level in the guild.

explicit_content_filter: ExplicitContentFilterLevel

The level at which explicit content will be filtered.

roles: HashMap<RoleId, Role>

The roles in the guild.

emojis: HashMap<EmojiId, Emoji>

The roles in the guild.

features: HashSet<GuildFeature>

The features enabled for the guild.

mfa_level: MfaLevel

The required MFA level for the guild.

application_id: Option<ApplicationId>

The application ID of the guild creator, if the guild is bot-created.

widget_enabled: bool

Whether the server widget is enabled.

widget_channel_id: Option<ChannelId>

The ID of the channel for the server widget.

system_channel_id: Option<ChannelId>

The ID of the channel to which system messages are sent.

joined_at: Option<DateTime<FixedOffset>>

When the guild was joined.

large: bool

Whether the guild is considered a large guild.

unavailable: bool

Whether the guild is unavailable.

member_count: Option<u64>

The total number of members in the guild.

voice_states: Vec<VoiceState>

The states of all current voice connections in the guild.

members: HashMap<UserId, Member>

The users in the guild.

channels: HashMap<ChannelId, GuildChannel>

The channels in the guild.

presences: HashMap<UserId, Presence>

The presences of the users in the guild.

max_members: Option<u64>

The maximum amount of members for the guild.

vanity_url_code: Option<String>

The vanity URL of the guild.

description: Option<String>

The description of the guild.

banner: Option<String>

The hash of the guild banner.

premium_tier: PremiumTier

The tier of premium for the guild, provided by Nitro boosts.

boost_count: u64

The total number of users currently boosting the guild.

preferred_locale: Locale

The preferred locale of the guild.

Defaults to en-US if the guild does not have the Discoverable feature enabled.